rename fap to nut
