+++ controllers/front/users/user.js
@@ -4,12 +4,12 @@ module.exports = async (req, res) => {
     var loggedInUser = await helpers.getLoggedInUser(req);
 
     //render results
-    const faps = await axios.get(`http://localhost:${req.app.locals.port}/api/user/${req.params.user}`);
+    const nuts = await axios.get(`http://localhost:${req.app.locals.port}/api/user/${req.params.user}`);
     
     res.render('pages/user', {
         loggedInUser: loggedInUser, 
         user: req.params.user,
-        faps: faps.data,
+        nuts: nuts.data,
         notifications: helpers.getNotifications(req)
     });
 }
