+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "fapcounter",
+  "name": "NuTracker",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
