+++ views/pages/modules/weekly-chart.pug
@@ -1,8 +1,8 @@
-mixin weekly-chart(faps)
+mixin weekly-chart(nuts)
     #weekly-chart
     script.
             /* Add a basic data series with six labels and values */
-            var faps = !{JSON.stringify(faps)};
+            var nuts = !{JSON.stringify(nuts)};
 
             var data = {
                 labels: ['6 days ago', '5 days ago', '4 days ago', '3 days ago', '2 days ago', '1 day ago', 'last 24h'],
@@ -17,8 +17,8 @@ mixin weekly-chart(faps)
                 }
             }
 
-            faps.forEach(fap => {
-                var daysPast = Math.floor((new Date().getTime() - fap.date) / (1000 * 3600 * 24));
+            nuts.forEach(nut => {
+                var daysPast = Math.floor((new Date().getTime() - nut.date) / (1000 * 3600 * 24));
                 
                 data.series[0].data[(data.series[0].data.length-1)-daysPast]++;
             });
