+++ views/pages/privacy.pug
@@ -8,7 +8,7 @@ body
     +nav
     .main-container
         .container
-            p We collect some information from you to be able to provide insight into your faps. Here is our policy:
+            p We collect some information from you to be able to provide insight into your nuts. Here is our policy:
                 ul
                     li.
                         We collect ALL of the data you give us, we save it in the database, 
@@ -18,6 +18,6 @@ body
                     li.
                         We may set cookies in your browser in order to keep your session open. When you log in, 
                         a cookie containing a session ID is saved in your browser, this cookie does not hold any login 
-                        information. it simply tells our servers that you're authenticated, so when you want to tell us you fapped,
+                        information. it simply tells our servers that you're authenticated, so when you want to tell us you nutted,
                         we can tell it's you, and we can check if it's you telling us or if you're authorised to tell us the person
-                        on who's behalf of you're fapping, has fapped.
+                        on who's behalf of you're nutting, has nutted.
